Add Strip type
A "Strip" type is introduced as a subclass of "Line" type. Most of the code is shared between the two types with some additional logic added to "Strip" to ensure the line is continuously drawn even if no new value is pushed to the drawing set. The type also forces specific drawing settings when "continuous drawing" is selected, namely the stepped drawing and the removal of any fill color.